Models changed, calculation algorithms added

+++ CommunalCalculatorTest/Models/Services/CommunalService.cs
@@ -3,6 +3,6 @@
     internal abstract class CommunalService
     {
         internal CommunalRate Rate { get; set; }
-        internal abstract double GetSalary();
+        internal abstract decimal GetSalary();
     }
 }
