+++ CommunalCalculatorTest/Models/Services/Imp/ColdWaterByNormative.cs
@@ -4,7 +4,7 @@
     {
         public ColdWaterByNormative(int residentsCount) : base(residentsCount) { }
 
-        internal override double GetSalary()
+        internal override decimal GetSalary()
         {
             throw new NotImplementedException();
         }
