+++ CommunalCalculatorTest/Models/Services/Imp/HeatCarrierByMeter.cs
@@ -4,7 +4,7 @@
     {
         public HeatCarrierByMeter(int scopeOfService) : base(scopeOfService) { }
         public HeatCarrierByMeter(int readingBefore, int readingNow) : base(readingBefore, readingNow) { }
-        internal override double GetSalary()
+        internal override decimal GetSalary()
         {
             throw new NotImplementedException();
         }
